Improve StoryTemplate component
